refactor(staking): consistent names for svg files
